:sparkles: Add SleepStage datatype. (#42)
'SleepStage' is a kind of 'IntervalDataType'. According to the decompiled code of 'com.google.android.libraries.healthdata:health-data-api:1.0.0-alpha01', it seems that it has 'stage' EnumField type field.
'''
    public static final class SleepStageDataType extends zzc implements IntervalDataType {
        @NonNull
        public final EnumField stage;

        SleepStageDataType() {
            super(SleepStage);
            EnumField var1 = new EnumField(stage, SleepStage.class, true, false);
            this.stage = var1;
            Field[] var2 = new Field[]{this.stage};
            this.initFields(var2);
        }
    }
'''
And that EnumField is set 'stage' key as some 'SleepStage.class' type of value.
The spec of 'SleepStage.class' is like below:
'''
public @interface SleepStage {
    @NonNull
    String UNKNOWN = unknown;
    @NonNull
    String AWAKE = awake;
    @NonNull
    String SLEEPING = sleeping;
    @NonNull
    String OUT_OF_BED = out_of_bed;
    @NonNull
    String LIGHT = light;
    @NonNull
    String DEEP = deep;
    @NonNull
    String REM = rem;
}
'''
So in summary, SleepStage has stage field as EnumFiled type. And that EnumField object has only 1 key-value pair. And that key-value pair is String-String typed pair.
The key is always 'stage' string, and the value is one of string value of 'SleepStage' type.

So in our platform, I think that it would be ok to set stage value as String type.

Signed-off-by: culater.kwak <culater.kwak@samsung.com>